+++ homework8/src/main/res/layout/fragment_main.xml
@@ -1,11 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity">
+    android:layout_height="match_parent">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolBar"
@@ -51,10 +48,10 @@
         app:layout_constraintTop_toBottomOf="@+id/toolBar" />
 
     <ImageView
-        android:scaleType="fitCenter"
         android:id="@+id/weatherMainImageView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:scaleType="fitCenter"
         android:src="@drawable/ic_ac_unit_black_40dp"
         app:layout_constraintBottom_toBottomOf="@id/cityTextView"
         app:layout_constraintStart_toEndOf="@id/cityTextView"
